调整商品详情显示页面3

+++ app/src/main/java/com/zhaidou/fragments/OrderAllOrdersFragment.java
@@ -392,7 +392,7 @@ public class OrderAllOrdersFragment extends BaseFragment implements View.OnClick
                             JSONObject orderObj = orderArr.optJSONObject(i);
                             int id = orderObj.optInt("id");
                             String number = orderObj.optString("number");
-                            int amount = orderObj.optInt("amount");
+                            double amount = orderObj.optDouble("amount");
                             String status = orderObj.optString("status");
                             String status_ch = orderObj.optString("status_ch");
                             String created_at = orderObj.optString("created_at");
